pyproject.toml
[tool.poetry]
name = "terraform-visualizer"
version = "0.1.0"
description = ""
authors = ["Joe Lombrozo <joe@djeebus.net>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
python-hcl2 = "^3.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

